Add diffScenarios function
This is the first in a (probable) series of functions for performing
commonly used transformations on GCAM data.  Some of these won't
exactly be GCAM-specific, inasmuch as they would work on non-GCAM
data, but they are/will be included because GCAM users frequently need
them, and because the defaults are selected with GCAM use cases in
mind.