Split scenario name and run date when query is retrieved from model interface.
The "scenario" column now contains just the scenario name.  The date
portion is automatically put into a column called "rundate".  This
happens in the cleanup at the end of runQuery, so in theory no table
should make it into the system without having this transformation
applied.  In practice we do some checks in later stages to make sure
that we don't get burned by legacy data that doesn't conform to this
convention.