Upgrade to BaseX 9 (#47)
* Upgrade to BaseX 9 and switch to it's new xquery serialization format
meant to deal with serializing/parsing large CSV datasets

* Update the modelinterface submodule to point to the `basex-9` branch on
the modelinterface (as apposed to master since we are upgraded rgcam
to BaseX 9 before GCAM).  Note the ModelInterface.jar also got rebuild,
this time with the minimum Java version as 8 instead of 7 as this is the
minimum in BaseX 9.  I've update the DESCRIPTION to reflect this.  Also
note we are implicitly getting fixes from the ModelInterface to support
running with Java 9/10.

* Update the vignette's BaseX server example on PIC to point to the
BaseX-9.0.1 install.

* Bump minor version since BaseX 9 can read older format databases however
BaseX 8 versions of rgcam won't be able to read BaseX 9 format databases

* Oops, just realized the vignette had the wrong path, src instead of lib

* Merge modelinterface master bugfixes; most particular to rgcam is the
fix when querying results that did not specify Units.